+++ src/main/kotlin/com/zepben/evolve/services/network/tracing/feeder/AssignToFeeders.kt
@@ -28,26 +28,46 @@ import com.zepben.evolve.services.network.tracing.traversal.StepContext
 class AssignToFeeders(
     internal val stateOperators: NetworkStateOperators
 ) {
-    fun run(network: NetworkService) {
-        val terminalToAuxEquipment = network.sequenceOf<AuxiliaryEquipment>()
-            .filter { it.terminal != null }
-            .groupBy { it.terminal!! }
 
-        val feederStartPoints = network.sequenceOf<Feeder>()
-            .mapNotNull { it.normalHeadTerminal?.conductingEquipment }
-            .toSet()
+    /**
+     * Assign equipment to feeders in the specified network, given an optional start terminal.
+     * * When a start terminal is provided, the trace will assign all feeders associated with the terminals equipment to all connected equipment.
+     * * If no start terminal is provided, all feeder head terminals in the network will be used instead, assigning their associated feeder.
+     */
+    @JvmOverloads
+    fun run(network: NetworkService, startTerminal: Terminal? = null) {
+        val feederStartPoints = network.feederStartPoints
+        val terminalToAuxEquipment = network.auxEquipmentByTerminal
 
-        network.sequenceOf<Feeder>().forEach { run(it, feederStartPoints, terminalToAuxEquipment) }
+        if (startTerminal == null) {
+            network.sequenceOf<Feeder>().forEach {
+                run(it.normalHeadTerminal, feederStartPoints, terminalToAuxEquipment, listOf(it))
+            }
+        } else
+            run(startTerminal, feederStartPoints, terminalToAuxEquipment, startTerminal.feeders(stateOperators))
     }
 
-    private fun run(
-        feeder: Feeder,
+    /**
+     * Assign equipment to feeders tracing out from the supplied terminal.
+     *
+     * @param terminal The [Terminal] to trace from.
+     * @param feederStartPoints A set of all [ConductingEquipment] containing a feeder head terminal. Must contain a list of all feeder
+     * head equipment that may be traced to from [terminal] to prevent excess tracing onto the sub-transmission network.
+     * @param terminalToAuxEquipment A map of all [AuxiliaryEquipment] by their attached [Terminal] that can be traced from [terminal].
+     * @param feedersToAssign The list of feeders to assign to all traced equipment.
+     */
+    fun run(
+        terminal: Terminal?,
         feederStartPoints: Set<ConductingEquipment>,
         terminalToAuxEquipment: Map<Terminal, List<AuxiliaryEquipment>>,
+        feedersToAssign: List<Feeder>,
     ) {
-        val headTerminal = feeder.normalHeadTerminal ?: return
-        val traversal = createTrace(terminalToAuxEquipment, feederStartPoints, listOf(feeder))
-        traversal.run(headTerminal, canStopOnStartItem = false)
+        // If there is no terminal, or there are no feeders assigned to the terminals equipment, then we have nothing to do.
+        if ((terminal == null) || (feedersToAssign.isEmpty()))
+            return
+
+        val traversal = createTrace(terminalToAuxEquipment, feederStartPoints, feedersToAssign)
+        traversal.run(terminal, canStopOnStartItem = false)
     }
 
     private fun createTrace(
@@ -96,3 +116,28 @@ class AssignToFeeders(
     }
 
 }
+
+/**
+ * A map of all [AuxiliaryEquipment] in the [NetworkService] indexed by their terminals.
+ */
+// TODO [Review]: Where should this be located?
+val NetworkService.auxEquipmentByTerminal: Map<Terminal, List<AuxiliaryEquipment>>
+    get() = sequenceOf<AuxiliaryEquipment>()
+        .filter { it.terminal != null }
+        .groupBy { it.terminal!! }
+
+/**
+ * A set of all [ConductingEquipment] in the [NetworkService] that are at the top of a feeder.
+ */
+// TODO [Review]: Where should this be located?
+val NetworkService.feederStartPoints: Set<ConductingEquipment>
+    get() = sequenceOf<Feeder>()
+        .mapNotNull { it.normalHeadTerminal?.conductingEquipment }
+        .toSet()
+
+/**
+ * Find the feeders assigned to the equipment of a terminal for the appropriate network state.
+ */
+// TODO [Review]: Where should this be located?
+fun Terminal.feeders(stateOperators: NetworkStateOperators): List<Feeder> =
+    conductingEquipment?.let { stateOperators.getContainers(it).filterIsInstance<Feeder>() }.orEmpty()