+++ src/test/kotlin/com/zepben/evolve/services/network/tracing/feeder/AssignToFeedersTest.kt
@@ -16,6 +16,7 @@ import com.zepben.evolve.cim.iec61970.base.core.Feeder
 import com.zepben.evolve.cim.iec61970.base.protection.CurrentRelay
 import com.zepben.evolve.cim.iec61970.base.protection.ProtectionRelayScheme
 import com.zepben.evolve.cim.iec61970.base.protection.ProtectionRelaySystem
+import com.zepben.evolve.cim.iec61970.base.wires.Junction
 import com.zepben.evolve.cim.iec61970.base.wires.ProtectedSwitch
 import com.zepben.evolve.services.network.testdata.*
 import com.zepben.evolve.services.network.tracing.networktrace.operators.NetworkStateOperators
@@ -202,6 +203,42 @@ internal class AssignToFeedersTest {
         AssignToFeeders(NetworkStateOperators.NORMAL).run(network)
 
         validateEquipment(feeder.equipment, "b0", "prsys3")
+    }
+
+    @Test
+    internal fun `can be run from a single terminal`() {
+        //
+        // b0 --c1-- j2 --c3-- --c4--
+        //           |
+        //           c5
+        //           |
+        //            --c6--
+        //
+        val network = TestNetworkBuilder()
+            .fromBreaker() // b0
+            .toAcls() // c1
+            .toJunction(numTerminals = 3) //j2
+            .toAcls() // c3
+            .toAcls() // c4
+            .fromAcls() // c5
+            .toAcls() // c6
+            .connect("j2", "c5", 2, 1)
+            .addFeeder("b0") // fdr7
+            .network
+
+        val feeder = network.get<Feeder>("fdr7")!!
+        val junction = network.get<Junction>("j2")!!.also {
+            feeder.addEquipment(it)
+            it.addContainer(feeder)
+        }
+
+        AssignToFeeders(NetworkStateOperators.NORMAL).run(network, startTerminal = junction.terminals.last())
+
+        // b0 is included from the network builder.
+        // j2 was added to allow us to test the terminal based assignment.
+        // c3 and c4 should have been added via the trace.
+        // c1, c5 and c6 shouldn't have been added if the assignment only went out t3 of j2.
+        validateEquipment(feeder.equipment, "b0", "j2", "c3", "c4")
     }
 
     private fun validateEquipment(equipment: Collection<Equipment>, vararg expectedMRIDs: String) {